@@ -29,8 +29,8 @@
         "khill/datatableplus": "Adds extra features to DataTables such as csv and laravel collection parsing."
     },
     "require": {
-        "php": ">=5.4.0",
-        "nesbot/carbon": ">=1.14 <4.0"
+        "php": ">=8.2",
+        "nesbot/carbon": ">=2.13 <4.0"
     },
     "autoload": {
         "psr-4": {
@@ -38,11 +38,11 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.8",
-        "mockery/mockery": "~0.9",
-        "squizlabs/php_codesniffer": "~2.5",
-        "satooshi/php-coveralls": "~1.0",
-        "codeclimate/php-test-reporter": "~0.3"
+        "roave/security-advisories": "dev-latest",
+        "phpunit/phpunit": "^11.0",
+        "mockery/mockery": "^1.6",
+        "squizlabs/php_codesniffer": "^3.9",
+        "php-coveralls/php-coveralls": "^2.2"
     },
     "autoload-dev": {
         "psr-4": {
@@ -71,7 +71,7 @@
     },
     "config": {
         "platform": {
-            "php": "5.4"
+            "php": "8.2"
         }
     },
     "minimum-stability": "stable"
