@@ -4,13 +4,17 @@ namespace Khill\Lavacharts\Tests\DataTables\Cells;
 
 use Khill\Lavacharts\Tests\ProvidersTestCase;
 use Khill\Lavacharts\DataTables\Cells\DateCell;
+use PHPUnit\Framework\Attributes\CoversMethod;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Depends;
 
+#[CoversMethod(DateCell::class, '__construct')]
+#[CoversMethod(DateCell::class, '__toString')]
+#[CoversMethod(DateCell::class, 'jsonSerialize')]
+#[CoversMethod(DateCell::class, 'parseString')]
 class DateCellTest extends ProvidersTestCase
 {
-    /**
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::__construct
-     */
-    public function testConstructorArgs()
+    public function testConstructorArgs(): void
     {
         $mockCarbon = \Mockery::mock('\Carbon\Carbon[parse]', ['2015-09-04 9:31:00']);
 
@@ -23,12 +27,7 @@ class DateCellTest extends ProvidersTestCase
         //$this->assertTrue(is_array($this->inspect($column, 'p')));
     }
 
-
-    /**
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::parseString
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::__toString
-     */
-    public function testParseStringWithNoFormat()
+    public function testParseStringWithNoFormat(): void
     {
         $cell = DateCell::parseString('3/24/1988 8:01:05');
         $this->assertEquals('Date(1988,2,24,8,1,5)', (string) $cell);
@@ -37,11 +36,7 @@ class DateCellTest extends ProvidersTestCase
         $this->assertEquals('Date(1988,2,24,8,1,5)', (string) $cell);
     }
 
-    /**
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::parseString
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::__toString
-     */
-    public function testParseStringWithFormat()
+    public function testParseStringWithFormat(): void
     {
 
         $cell = DateCell::parseString('5:45pm on Saturday 24th March 2012', 'g:ia \o\n l jS F Y');
@@ -49,48 +44,33 @@ class DateCellTest extends ProvidersTestCase
         $this->assertEquals('Date(2012,2,24,17,45,0)', (string) $cell);
     }
 
-    /**
-     * @expectedException \Exception
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::parseString
-     */
-    public function testParseStringWithBadDateTimeString()
+    public function testParseStringWithBadDateTimeString(): void
     {
+        $this->expectException(\Exception::class);
         DateCell::parseString('132/06/199210');
     }
 
-    /**
-     * @dataProvider nonStringOrNullProvider
-     * @expectedException \Khill\Lavacharts\Exceptions\InvalidDateTimeString
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::parseString
-     */
-    public function testParseStringWithBadTypesForDateTime($badTypes)
+    #[DataProvider('nonStringOrNullProvider')]
+    public function testParseStringWithBadTypesForDateTime($badTypes): void
     {
+        $this->expectException(\Khill\Lavacharts\Exceptions\InvalidDateTimeString::class);
         DateCell::parseString($badTypes);
     }
 
-    /**
-     * @expectedException \Khill\Lavacharts\Exceptions\InvalidDateTimeFormat
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::parseString
-     */
-    public function testParseStringWithBadFormatString()
+    public function testParseStringWithBadFormatString(): void
     {
+        $this->expectException(\Khill\Lavacharts\Exceptions\InvalidDateTimeFormat::class);
         DateCell::parseString('1/2/2003', 'sushi');
     }
 
-    /**
-     * @dataProvider nonStringProvider
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::parseString
-     */
-    public function testParseStringWithBadTypesForFormat()
+    #[DataProvider('nonStringProvider')]
+    public function testParseStringWithBadTypesForFormat(): void
     {
-        DateCell::parseString('1/2/2003', ['imnotaformat']);
+        self::assertInstanceOf(DateCell::class, DateCell::parseString('1/2/2003', ['imnotaformat']));
     }
 
-    /**
-     * @depends testConstructorArgs
-     * @covers \Khill\Lavacharts\DataTables\Cells\DateCell::jsonSerialize
-     */
-    public function testJsonSerialization()
+    #[Depends('testConstructorArgs')]
+    public function testJsonSerialization(): void
     {
         $mockCarbon = \Mockery::mock('\Carbon\Carbon[parse]', ['2015-09-04 9:31:00']);
 
