@@ -62,20 +62,18 @@ class LavachartsServiceProvider extends ServiceProvider
 
         $defaultConfig = $this->app['config']->get('lavacharts');
 
-        $this->app->singleton('lavacharts', function() use ($defaultConfig) {
+        $this->app->singleton('lavacharts', function () use ($defaultConfig) {
             return new Lavacharts($defaultConfig);
         });
 
-        $this->app->booting(function() {
+        $this->app->booting(function () {
             $loader = AliasLoader::getInstance();
             $loader->alias('Lava', 'Khill\Lavacharts\Laravel\LavachartsFacade');
         });
-
     }
 
     public function provides()
     {
         return ['lavacharts'];
     }
-
 }
