@@ -4,7 +4,6 @@ namespace Khill\Lavacharts;
 
 use Khill\Lavacharts\Charts\Chart;
 use Khill\Lavacharts\Charts\ChartFactory;
-use Khill\Lavacharts\Dashboards\Dashboard;
 use Khill\Lavacharts\Dashboards\DashboardFactory;
 use Khill\Lavacharts\Dashboards\Filters\Filter;
 use Khill\Lavacharts\Dashboards\Filters\FilterFactory;
@@ -17,27 +16,25 @@ use Khill\Lavacharts\Exceptions\InvalidLabel;
 use Khill\Lavacharts\Exceptions\InvalidLavaObject;
 use Khill\Lavacharts\Javascript\ScriptManager;
 use Khill\Lavacharts\Support\Config;
+use Khill\Lavacharts\Support\Contracts\RenderableInterface as Renderable;
 use Khill\Lavacharts\Support\Html\HtmlFactory;
 use Khill\Lavacharts\Support\Psr4Autoloader;
+use Khill\Lavacharts\Support\Traits\HasOptionsTrait as HasOptions;
 use Khill\Lavacharts\Values\ElementId;
 use Khill\Lavacharts\Values\Label;
 use Khill\Lavacharts\Values\StringValue;
-use Khill\Lavacharts\Support\Traits\HasOptionsTrait as HasOptions;
-use Khill\Lavacharts\Support\Contracts\RenderableInterface as Renderable;
-
-require(__DIR__.'/Support/Traits/HasOptionsTrait.php');
 
 /**
  * Lavacharts - A PHP wrapper library for the Google Chart API
  *
  *
- * @category  Class
- * @package   Khill\Lavacharts
- * @author    Kevin Hill <kevinkhill@gmail.com>
+ * @category      Class
+ * @package       Khill\Lavacharts
+ * @author        Kevin Hill <kevinkhill@gmail.com>
  * @copyright (c) 2017, KHill Designs
- * @link      http://github.com/kevinkhill/lavacharts GitHub Repository Page
- * @link      http://lavacharts.com                   Official Docs Site
- * @license   http://opensource.org/licenses/MIT      MIT
+ * @link          http://github.com/kevinkhill/lavacharts GitHub Repository Page
+ * @link          http://lavacharts.com                   Official Docs Site
+ * @license       http://opensource.org/licenses/MIT      MIT
  */
 class Lavacharts
 {
@@ -82,7 +79,7 @@ class Lavacharts
      */
     public function __construct(array $options = [])
     {
-        if ( ! $this->usingComposer()) {
+        if (!$this->usingComposer()) {
             require_once(__DIR__.'/Support/Psr4Autoloader.php');
 
             $loader = new Psr4Autoloader;
@@ -92,22 +89,23 @@ class Lavacharts
 
         $this->initializeOptions($options);
 
-        $this->volcano       = new Volcano;
-        $this->chartFactory  = new ChartFactory;
-        $this->dashFactory   = new DashboardFactory;
+        $this->volcano = new Volcano;
+        $this->chartFactory = new ChartFactory;
+        $this->dashFactory = new DashboardFactory;
         $this->scriptManager = new ScriptManager($this->options);
     }
 
     /**
      * Magic function to reduce repetitive coding and create aliases.
      *
-     * @since  1.0.0
-     * @param  string $method Name of method
-     * @param  array  $args   Passed arguments
-     * @throws \Khill\Lavacharts\Exceptions\InvalidLabel
+     * @param string $method Name of method
+     * @param array  $args   Passed arguments
+     *
+     * @return mixed Returns Charts, Formats and Filters
      * @throws \Khill\Lavacharts\Exceptions\InvalidLavaObject
      * @throws \Khill\Lavacharts\Exceptions\InvalidFunctionParam
-     * @return mixed Returns Charts, Formats and Filters
+     * @throws \Khill\Lavacharts\Exceptions\InvalidLabel
+     * @since  1.0.0
      */
     public function __call($method, $args)
     {
@@ -142,7 +140,7 @@ class Lavacharts
             $lavaClass = Format::create($method, $options);
         }
 
-        if (isset($lavaClass) == false) {
+        if (!isset($lavaClass)) {
             throw new InvalidLavaObject($method);
         }
 
@@ -152,8 +150,8 @@ class Lavacharts
     /**
      * Get the ScriptManager instance.
      *
-     * @since 3.1.9
      * @return ScriptManager
+     * @since 3.1.9
      */
     public function getScriptManager()
     {
@@ -166,12 +164,13 @@ class Lavacharts
      * If the additional DataTablePlus package is available, then one will
      * be created, otherwise a standard DataTable is returned.
      *
+     * @param mixed $args
+     *
+     * @return \Khill\Lavacharts\DataTables\DataTable
      * @since  3.0.3
      * @uses   \Khill\Lavacharts\DataTables\DataFactory
-     * @param  mixed $args
-     * @return \Khill\Lavacharts\DataTables\DataTable
      */
-    public function DataTable($args = null)
+    public function DataTable($args = null): DataTable // phpcs:ignore PSR1.Methods.CamelCapsMethodName
     {
         $dataFactory = __NAMESPACE__.'\\DataTables\\DataFactory::DataTable';
 
@@ -181,12 +180,13 @@ class Lavacharts
     /**
      * Create a new Dashboard
      *
-     * @since  3.0.0
-     * @param  string                                 $label
-     * @param  \Khill\Lavacharts\DataTables\DataTable $dataTable
+     * @param string                                 $label
+     * @param \Khill\Lavacharts\DataTables\DataTable $dataTable
+     *
      * @return \Khill\Lavacharts\Dashboards\Dashboard
+     * @since  3.0.0
      */
-    public function Dashboard($label, DataTable $dataTable)
+    public function Dashboard($label, DataTable $dataTable): Dashboards\Dashboard // phpcs:ignore PSR1.Methods.CamelCapsMethodName
     {
         if ($this->exists('Dashboard', $label)) {
             $dashboard = $this->volcano->get('Dashboard', $label);
@@ -202,13 +202,14 @@ class Lavacharts
     /**
      * Create a new ControlWrapper from a Filter
      *
-     * @since  3.0.0
-     * @uses   \Khill\Lavacharts\Values\ElementId
-     * @param  \Khill\Lavacharts\Dashboards\Filters\Filter $filter Filter to wrap
-     * @param  string $elementId HTML element ID to output the control.
+     * @param \Khill\Lavacharts\Dashboards\Filters\Filter $filter    Filter to wrap
+     * @param string                                      $elementId HTML element ID to output the control.
+     *
      * @return \Khill\Lavacharts\Dashboards\Wrappers\ControlWrapper
+     * @uses   \Khill\Lavacharts\Values\ElementId
+     * @since  3.0.0
      */
-    public function ControlWrapper(Filter $filter, $elementId)
+    public function ControlWrapper(Filter $filter, $elementId): ControlWrapper // phpcs:ignore PSR1.Methods.CamelCapsMethodName
     {
         $elementId = new ElementId($elementId);
 
@@ -218,13 +219,14 @@ class Lavacharts
     /**
      * Create a new ChartWrapper from a Chart
      *
-     * @since  3.0.0
-     * @uses   \Khill\Lavacharts\Values\ElementId
-     * @param  \Khill\Lavacharts\Charts\Chart $chart Chart to wrap
-     * @param  string $elementId HTML element ID to output the control.
+     * @param \Khill\Lavacharts\Charts\Chart $chart     Chart to wrap
+     * @param string                         $elementId HTML element ID to output the control.
+     *
      * @return \Khill\Lavacharts\Dashboards\Wrappers\ChartWrapper
+     * @uses   \Khill\Lavacharts\Values\ElementId
+     * @since  3.0.0
      */
-    public function ChartWrapper(Chart $chart, $elementId)
+    public function ChartWrapper(Chart $chart, $elementId): ChartWrapper // phpcs:ignore PSR1.Methods.CamelCapsMethodName
     {
         $elementId = new ElementId($elementId);
 
@@ -239,10 +241,11 @@ class Lavacharts
      * By default, Lavacharts is loaded with the "en" locale. You can override this default
      * by explicitly specifying a locale when creating the DataTable.
      *
-     * @since  3.1.0
-     * @param  string $locale
+     * @param string $locale
+     *
      * @return $this
      * @throws \Khill\Lavacharts\Exceptions\InvalidStringValue
+     * @since  3.1.0
      */
     public function setLocale($locale = 'en')
     {
@@ -250,11 +253,12 @@ class Lavacharts
 
         return $this;
     }
+
     /**
      * Returns the current locale used in the DataTable
      *
-     * @since  3.1.0
      * @return string
+     * @since  3.1.0
      */
     public function getLocale()
     {
@@ -266,13 +270,13 @@ class Lavacharts
      *
      * Will be depreciating jsapi in the future
      *
-     * @since  3.0.3
      * @return string Google Chart API and lava.js script blocks
+     * @since  3.0.3
      */
     public function lavajs()
     {
         $config = [
-            'locale' => $this->locale
+            'locale' => $this->locale,
         ];
 
         return (string) $this->scriptManager->getLavaJsModule($config);
@@ -281,9 +285,9 @@ class Lavacharts
     /**
      * Outputs the link to the Google JSAPI
      *
-     * @since      2.3.0
-     * @deprecated 3.0.3
      * @return string Google Chart API and lava.js script blocks
+     * @deprecated 3.0.3
+     * @since      2.3.0
      */
     public function jsapi()
     {
@@ -293,11 +297,12 @@ class Lavacharts
     /**
      * Checks to see if the given chart or dashboard exists in the volcano storage.
      *
-     * @since  2.4.2
-     * @uses   \Khill\Lavacharts\Values\Label
-     * @param  string $type Type of object to isNonEmpty.
-     * @param  string $label Label of the object to isNonEmpty.
+     * @param string $type  Type of object to isNonEmpty.
+     * @param string $label Label of the object to isNonEmpty.
+     *
      * @return boolean
+     * @uses   \Khill\Lavacharts\Values\Label
+     * @since  2.4.2
      */
     public function exists($type, $label)
     {
@@ -313,12 +318,13 @@ class Lavacharts
     /**
      * Fetches an existing Chart or Dashboard from the volcano storage.
      *
-     * @since  3.0.0
-     * @uses   \Khill\Lavacharts\Values\Label
-     * @param  string $type  Type of Chart or Dashboard.
-     * @param  string $label Label of the Chart or Dashboard.
+     * @param string $type  Type of Chart or Dashboard.
+     * @param string $label Label of the Chart or Dashboard.
+     *
      * @return \Khill\Lavacharts\Support\Contracts\RenderableInterface
      * @throws \Khill\Lavacharts\Exceptions\InvalidLavaObject
+     * @since  3.0.0
+     * @uses   \Khill\Lavacharts\Values\Label
      */
     public function fetch($type, $label)
     {
@@ -334,9 +340,10 @@ class Lavacharts
     /**
      * Stores a existing Chart or Dashboard into the volcano storage.
      *
-     * @since  3.0.0
-     * @param  \Khill\Lavacharts\Support\Contracts\RenderableInterface $renderable A Chart or Dashboard.
+     * @param \Khill\Lavacharts\Support\Contracts\RenderableInterface $renderable A Chart or Dashboard.
+     *
      * @return \Khill\Lavacharts\Support\Contracts\RenderableInterface
+     * @since  3.0.0
      */
     public function store(Renderable $renderable)
     {
@@ -352,15 +359,16 @@ class Lavacharts
      * As of version 3.1, the elementId parameter is optional, but only
      * if the elementId was set explicitly to the Renderable.
      *
-     * @since  2.0.0
+     * @param string $type      Type of object to render.
+     * @param string $label     Label of the object to render.
+     * @param mixed  $elementId HTML element id to render into.
+     * @param mixed  $div       Set true for div creation, or pass an array with height & width
+     *
+     * @return string
      * @uses   \Khill\Lavacharts\Values\Label
      * @uses   \Khill\Lavacharts\Values\ElementId
      * @uses   \Khill\Lavacharts\Support\Buffer
-     * @param  string $type       Type of object to render.
-     * @param  string $label      Label of the object to render.
-     * @param  mixed  $elementId  HTML element id to render into.
-     * @param  mixed  $div        Set true for div creation, or pass an array with height & width
-     * @return string
+     * @since  2.0.0
      */
     public function render($type, $label, $elementId = null, $div = false)
     {
@@ -388,8 +396,8 @@ class Lavacharts
     /**
      * Renders all charts and dashboards that have been defined
      *
-     * @since  3.1.0
      * @return string
+     * @since  3.1.0
      */
     public function renderAll()
     {
@@ -414,15 +422,16 @@ class Lavacharts
      * Given a chart label and an HTML element id, this will output
      * all of the necessary javascript to generate the chart.
      *
-     * @since  3.0.0
-     * @param  string                             $type
-     * @param  \Khill\Lavacharts\Values\Label     $label
-     * @param  \Khill\Lavacharts\Values\ElementId $elementId HTML element id to render the chart into.
-     * @param  bool|array                         $div       Set true for div creation, or pass an array with height & width
+     * @param string                             $type
+     * @param \Khill\Lavacharts\Values\Label     $label
+     * @param \Khill\Lavacharts\Values\ElementId $elementId HTML element id to render the chart into.
+     * @param bool|array                         $div       Set true for div creation, or pass an array with height & width
+     *
      * @return \Khill\Lavacharts\Support\Buffer
      * @throws \Khill\Lavacharts\Exceptions\ChartNotFound
      * @throws \Khill\Lavacharts\Exceptions\InvalidConfigValue
      * @throws \Khill\Lavacharts\Exceptions\InvalidDivDimensions
+     * @since  3.0.0
      */
     private function renderChart($type, Label $label, ElementId $elementId = null, $div = false)
     {
@@ -455,12 +464,13 @@ class Lavacharts
      * Given a chart label and an HTML element id, this will output
      * all of the necessary javascript to generate the chart.
      *
-     * @since  3.0.0
-     * @uses   \Khill\Lavacharts\Support\Buffer   $buffer
-     * @param  \Khill\Lavacharts\Values\Label     $label
-     * @param  \Khill\Lavacharts\Values\ElementId $elementId HTML element id to render the chart into.
+     * @param \Khill\Lavacharts\Values\Label     $label
+     * @param \Khill\Lavacharts\Values\ElementId $elementId HTML element id to render the chart into.
+     *
      * @return \Khill\Lavacharts\Support\Buffer
      * @throws \Khill\Lavacharts\Exceptions\DashboardNotFound
+     * @since  3.0.0
+     * @uses   \Khill\Lavacharts\Support\Buffer   $buffer
      */
     private function renderDashboard(Label $label, ElementId $elementId = null)
     {
@@ -486,8 +496,8 @@ class Lavacharts
      * This will check if the folder 'composer' is within the path to Lavacharts.
      *
      * @access private
-     * @since  2.4.0
      * @return boolean
+     * @since  2.4.0
      */
     private function usingComposer()
     {
@@ -503,6 +513,7 @@ class Lavacharts
      * passed values.
      *
      * @param array $options
+     *
      * @return void
      */
     private function initializeOptions(array $options)
