@@ -4,14 +4,24 @@ namespace Khill\Lavacharts\Tests\DataTables\Columns;
 
 use Khill\Lavacharts\Tests\ProvidersTestCase;
 use Khill\Lavacharts\DataTables\Columns\Column;
+use PHPUnit\Framework\Attributes\CoversMethod;
+use PHPUnit\Framework\Attributes\Depends;
 
 /**
  * @property \Mockery\Mock mockRole
  * @property \Mockery\Mock mockFormat
  */
+#[CoversMethod(Column::class, '__construct')]
+#[CoversMethod(Column::class, 'getType')]
+#[CoversMethod(Column::class, 'getLabel')]
+#[CoversMethod(Column::class, 'isFormatted')]
+#[CoversMethod(Column::class, 'getFormat')]
+#[CoversMethod(Column::class, 'getRole')]
+#[CoversMethod(Column::class, 'jsonSerialize')]
+#[\AllowDynamicProperties]
 class ColumnTest extends ProvidersTestCase
 {
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
@@ -24,107 +34,77 @@ class ColumnTest extends ProvidersTestCase
         $this->mockFormat = \Mockery::mock('\Khill\Lavacharts\DataTables\Formats\NumberFormat')->makePartial();
     }
 
-    /**
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::__construct
-     */
-    public function testConstructorWithType()
+    public function testConstructorWithType(): void
     {
         $column = new Column('number');
 
         $this->assertEquals('number', $this->inspect($column, 'type'));
     }
 
-    /**
-     * @depends testConstructorWithType
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::__construct
-     */
-    public function testConstructorWithTypeAndLabel()
+    #[Depends('testConstructorWithType')]
+    public function testConstructorWithTypeAndLabel(): void
     {
         $column = new Column('number', 'MyLabel');
 
         $this->assertEquals('MyLabel', $this->inspect($column, 'label'));
     }
 
-    /**
-     * @depends testConstructorWithTypeAndLabel
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::__construct
-     */
-    public function testConstructorWithTypeAndLabelAndFormat()
+    #[Depends('testConstructorWithTypeAndLabel')]
+    public function testConstructorWithTypeAndLabelAndFormat(): void
     {
         $column = new Column('number', 'MyLabel', $this->mockFormat);
 
         $this->assertInstanceOf('\Khill\Lavacharts\DataTables\Formats\NumberFormat', $this->inspect($column, 'format'));
     }
 
-    /**
-     * @depends testConstructorWithTypeAndLabelAndFormat
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::__construct
-     */
-    public function testConstructorWithTypeAndLabelAndFormatAndRole()
+    #[Depends('testConstructorWithTypeAndLabelAndFormat')]
+    public function testConstructorWithTypeAndLabelAndFormatAndRole(): void
     {
         $column = new Column('number', 'MyLabel', $this->mockFormat, $this->mockRole);
 
         $this->assertInstanceOf('\Khill\Lavacharts\Values\Role', $this->inspect($column, 'role'));
     }
 
-    /**
-     * @depends testConstructorWithType
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::getType
-     */
-    public function testGetType()
+    #[Depends('testConstructorWithType')]
+    public function testGetType(): void
     {
         $column = new Column('number');
 
         $this->assertEquals('number', $column->getType());
     }
 
-    /**
-     * @depends testConstructorWithTypeAndLabel
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::getLabel
-     */
-    public function testGetLabel()
+    #[Depends('testConstructorWithTypeAndLabel')]
+    public function testGetLabel(): void
     {
         $column = new Column('number', 'MyLabel');
 
         $this->assertEquals('MyLabel', $column->getLabel());
     }
 
-    /**
-     * @depends testConstructorWithTypeAndLabelAndFormat
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::isFormatted
-     */
-    public function testIsFormatted()
+    #[Depends('testConstructorWithTypeAndLabelAndFormat')]
+    public function testIsFormatted(): void
     {
         $column = new Column('number', 'MyLabel', $this->mockFormat);
 
         $this->assertTrue($column->isFormatted());
     }
 
-    /**
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::getFormat
-     */
-    public function testGetFormat()
+    public function testGetFormat(): void
     {
         $column = new Column('number', 'MyLabel', $this->mockFormat);
 
         $this->assertInstanceOf('\Khill\Lavacharts\DataTables\Formats\NumberFormat', $column->getFormat());
     }
 
-    /**
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::getRole
-     */
-    public function testGetRole()
+    public function testGetRole(): void
     {
         $column = new Column('number', 'MyLabel', $this->mockFormat, $this->mockRole);
 
         $this->assertEquals('interval', $column->getRole());
     }
 
-    /**
-     * @depends testConstructorWithTypeAndLabelAndFormatAndRole
-     * @covers \Khill\Lavacharts\DataTables\Columns\Column::jsonSerialize
-     */
-    public function testJsonSerialization()
+    #[Depends('testConstructorWithTypeAndLabelAndFormatAndRole')]
+    public function testJsonSerialization(): void
     {
         $column = new Column('number', 'MyLabel', $this->mockFormat, $this->mockRole);
 
@@ -133,6 +113,3 @@ class ColumnTest extends ProvidersTestCase
         $this->assertEquals($json, json_encode($column));
     }
 }
-
-
-
