@@ -4,9 +4,9 @@
          ->addNumberColumn('Thing 1')
          ->addNumberColumn('Thing 2');
 
-    for ($a=1;$a<30;$a++) {
-        $data->addRow(['2017-1-'.$a, rand(100,200), rand(100,200)]);
-    }
+for ($a=1; $a<30; $a++) {
+    $data->addRow(['2017-1-'.$a, rand(100, 200), rand(100, 200)]);
+}
 
     $lava->AreaChart($title, $data, [
         'width' => $width,
