@@ -3,9 +3,9 @@
     $data->addNumberColumn('Age')
          ->addNumberColumn('Weight');
 
-    for ($i=0; $i < 50; $i++) {
-        $data->addRow([rand(20, 40), rand(100, 300)]);
-    }
+for ($i=0; $i < 50; $i++) {
+    $data->addRow([rand(20, 40), rand(100, 300)]);
+}
 
     $lava->ScatterChart($title, $data, [
         'title' => 'Age vs. Weight comparison',
